+++ CHANGELOG.md
@@ -9,6 +9,7 @@ mod).
 
 ### Fixed
 
+- Fixed Puncture logic: Now the damage is dealth before Vulnerable is applied
 - Fixed the card library shifting its cards off center when a search matched only Mixes
 
 ## [0.14.16] - 2026-09-21